pyproject.toml
[tool.poetry]
name = "welbex"
version = "0.1.0"
description = ""
authors = ["Rail <rail.akhmietshin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
asyncpg = "^0.27.0"
alembic = "^1.11.1"
sqlalchemy = "^2.0.15"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
redis = "^4.5.5"
celery = "^5.3.0"
flower = "^1.2.0"
geopy = "^2.3.0"



[tool.poetry.dev-dependencies]
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
asyncpg = "^0.27.0"
alembic = "^1.11.1"
SQLAlchemy = "^2.0.15"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
geopy = "^2.3.0"
celery = "^5.3.0"
flower = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
